requirements.txt
torch>=2.2.0
transformers>=4.38.0
datasets>=2.18.0
pandas>=2.2.0
scikit-learn>=1.4.0
spacy>=3.7.0
wikipedia-api>=0.6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
jupyter>=1.0.0
accelerate>=0.27.0